requirements.txt
flask
flask-cors
sqlalchemy
paho-mqtt
